+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 
 ## [Unreleased]
 
+### Changed
+
+- Added timezone-aware dates and times to startup, access, error, and panic trace logs.
+
 ## [0.2.22] - 2026-07-29
 
 ### Changed